# backend/src/main/resources/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Pickleball Tournament Management System API
  description: |
    RESTful API for managing pickleball tournaments, events, matches, and user authentication.
    This specification is derived from the frontend client requirements.
  version: 1.0.0
  contact:
    name: PTMS Development Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/signup
  schemas:
...
    Event:
      type: object
      required:
        - id
        - name
        - tournamentIds
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Summer Pickleball Championship 2026"
        description:
          type: string
          nullable: true
          example: "Annual summer tournament series"
        startDate:
          type: string
          format: date
          nullable: true
          example: "2026-06-01"
        endDate:
          type: string
          format: date
          nullable: true
          example: "2026-06-15"
        tournamentIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true

    # ----------------------------------------------------------
    # Tournament Schemas
    # ----------------------------------------------------------
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        error:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          example: "Resource not found"

# ============================================================
# PATHS
# ============================================================
paths:
...
  /events/{eventId}:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Events
      summary: Get event by ID
      operationId: getEventById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Events
      summary: Update an event
      operationId: updateEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Events
      summary: Delete an event (and all tournaments within)
      operationId: deleteEvent
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...